feat: add Video attachment component as a wrapper around Media component
